refactor: extract shared logging to @flaim/worker-shared
All 5 workers had identical logEvalEvent() implementations differing only
in their TraceLogEvent interface fields. Extract a unified superset
interface and function to workers/shared/src/logging.ts, replace each
worker's logging.ts with a thin re-export.

Also adds @flaim/worker-shared as a dependency of auth-worker and aligns
its tsconfig with the other workers (moduleResolution: bundler, path
aliases, ESNext module target).
